An R6 class to assess a model with a scalar metric

Assess how well a model can predict time to event less than a certain threshold with a scalar metric (like the ROC-AUC).

Public fields

metrics

Assess the model for these scalar metrics. Check out the initializer for possible choices.

pivot_time_cutoff

Assess classifying time to event less than pivot_time_cutoff.

benchmark

Incorporate the benchmark into the assessment.

round_digits

Round the results in tables to round_digits digits after the point.

file

Save the resulting tibble to this csv file. This is the file for the train cohort, we infer the the file name of the test cohort by mirroring this file name. See the assess_center() method for more details.

Methods

Public methods

AssScalar$new()

AssScalar$assess()

AssScalar$assess_center()

AssScalar$clone()

Method new()

Create a new AssScalar instance.

Usage

AssScalar$new(
  metrics = c("auc", "accuracy", "precision", "prevalence", "n_true", "perc_true",
    "n_samples", "logrank", "threshold"),
  pivot_time_cutoff,
  benchmark = NULL,
  file = NULL,
  round_digits = 3
)

Arguments

metrics

character. Assess the model for these metrics. For currently offered choices see "Usage" above. If you have a model with non-binary output (like a logistic regression), we choose a threshold by maximizing the left-most metric in metrics that is made for classifiers with binary output (often this is "accuracy"). If this cannot be done reasonably (e.g., "precision" would usually give you a very high threshold with very low prevalence), we throw an error.

pivot_time_cutoff

numeric. Assess for classifying time to event < pivot_time_cutoff.

benchmark

character or NULL. If not NULL, include benchmark (the name of column in the pheno data) in the assessment.

file

string or NULL. The name of the csv file to save the results to for the train cohort. the results are not saved.

round_digits

numeric. The number of digits to round the results to. Default is 3.

Returns

A new AssScalar object.

Method assess()

Assess a single model (with multiple splits) on a data set.

Usage

AssScalar$assess(data, model, quiet = FALSE)

Arguments

data

Data object. Assess on this data. Data must already be read in and cohort attribute set.

model

Model object. Assess this model, multiple splits are supported.

quiet

logical. Whether to suppress messages.

Details

The AssScalar S3 class is tailored for this function.

Returns

numeric matrix. The rows correspond to the splits of the model, the columns correspond to self$metrics.

Method assess_center()

Assess multiple models (with multiple splits) on a data set.

Usage

AssScalar$assess_center(
  data,
  model_list,
  mirror = c("models", "results"),
  quiet = FALSE
)

Arguments

data

Data object. Assess on this data. If data$cohort is NULL, assess on the test cohort.

model_list

list of Model objects. Assess these models.

mirror

character vector of length 2. If you want to store the resulting tibbles, get the test-cohort tibble's file name by mirroring ass_scalar$file according to model_tree_mirror, i.e. replacing model_tree_mirror[1] by model_tree_mirror[2] in the file attribute.

quiet

logical. Whether to suppress messages.

Returns

A tibble. Every row stands for one model.

If length(self$metrics) == 1, the columns hold statistics on the single metric (like mean, standard deviation).

Otherwise columns correspond to self$metrics.

Method clone()

The objects of this class are cloneable with this method.

Usage

AssScalar$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.