An R6 class for a model

A Model specifies how a model looks like, prepares the data, fit_obj it, stores it and predicts from it.

See also

Public fields

name

A telling name for the model.

directory

Store/find the models in this directory.

fitter

The model fitting function to be used.

time_cutoffs

Threshold and censor the outcome accordingly.

val_error_fun

Calculate the error of independently validated predictions.

hyperparams

Optional arguments passed to fitter.

include_from_continuous_pheno

The names of the continuous variables in the pheno data (to be) included in the predictor matrix.

include_from_discrete_pheno

The names of the discrete variables in the pheno data (to be) included in the predictor matrix.

include_expr

Whether to include the expression data in the predictor matrix.

li_var_suffix

Append this to the names of features from the pheno data when adding them to the predictor matrix.

create_directory

Whether to create directory if it does not exist, yet.

file

Store this Model object under this name in directory.

fit_obj

A list holding fit_obj (something returned by a fitter like a ptk_zerosum S3 object).

continuous_output

Whether the output of the model is continous or binary.

combine_n_max_categorical_features

Maximum number of categorical features to combine

combined_feature_min_positive_ratio

Minimum ratio of positive observations in a combined (categorical) feature

Methods

Public methods

Model$new()

Model$fit()

Model$predict()

Model$clone()

Method new()

Create a new Model instance.

Usage

Model$new(
  name,
  fitter,
  directory,
  time_cutoffs,
  val_error_fun,
  hyperparams = NULL,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  include_expr = TRUE,
  li_var_suffix = "++",
  create_directory = TRUE,
  file = "model.rds",
  continuous_output = NULL,
  combine_n_max_categorical_features = 1L,
  combined_feature_min_positive_ratio = 0.04
)

Arguments

name

string. A telling name for the model.

fitter

function. The model fitting function to be used. Must take x and y as first two positional arguments. Further arguments can be passed via hyperparams (below). Its return value must be an S3 object with a plot() method, and (ideally, for assessment) with a predict() method. Default is NULL.

directory

string. The directory to store the models in. For every value in time_cutoffs, find the corresponding model in a subdirectory named after this value.

time_cutoffs

numeric vector. This governs how the provided response looks like. For every value of time_cutoffs, specify a model on the following response data:

For Cox response, censor all patients where the event ouccured after time_cutoffs at this value and train the specified model.

For binary response, binarize the outcome depending on whether it occured before or after this value.

When fitting models to the data, you find every of the length(time_cutoffs) models in a subdirectory of directory named after the respective value in time_cutoffs.

val_error_fun

Function used to calculate the error of independently validated predictions. Must take two numeric vector of equal length: y and y_hat, the true and predicted outcomes, respectively, and return a numeric scalar; the lower, the better the model. See error_rate() or neg_roc_auc() for examples.

hyperparams

list. Optional arguments passed to fitter, e.g. alpha in case of an elastic net. Default is list(), i.e., no arguments other than x, y passed to fitter.

include_from_continuous_pheno

vector of strings. The names of the continuous variables in the pheno data (to be) included in the predictor matrix. The values will be coerced to numeric. Default is NULL, which means no continuous pheno variables are or will be included.

include_from_discrete_pheno

vector of strings. The names of the discrete variables in the pheno data (to be) included in the predictor matrix. A discrete variable with n levels will be converted to n-1 binary variables. Default is NULL, which means no discrete pheno variables are or will be included.

include_expr

logical. Whether to include the expression data in the predictor matrix.

li_var_suffix

string. Append this to the names of features from the pheno data when adding them to the predictor matrix. Default is "++".

create_directory

logical. Whether to create directory if it does not exist, yet. Default is TRUE.

file

string. The name of the model-fit_obj file inside directory. Default is "fit_obj.rds".

continuous_output

logical or NULL. Whether the output of the model is continous (like the conditional probabilities of a logistic regression) or binary (like with a random forest). This piece of information is helpful when it comes to assessing a model (Ass2d$assess_center() uses it, e.g.).

combine_n_max_categorical_features

integer. Maximum number of categorical features to combine in predicting features.

combined_feature_min_positive_ratio

numeric. Minimum ratio of positive observations in a combined (categorical) feature. This attribute together with combine_n_max_categorical_features governs which combinatorial features the predicitor matrix will contain.

Returns

A Model R6 object.

Method fit()

Fit the model to a data set.

Usage

Model$fit(data, quiet = FALSE, msg_prefix = "")

Arguments

data

Data object. Read it in if needed.

quiet

logical. Whether to suppress messages. Default is FALSE.

msg_prefix

string. Prefix for messages. Default is "".

Returns

The Model object itself with the fit_obj attribute set to the object tuned over time_cutoffs, combine_n_max_categorical_features and hyperparams with fitter at the core.

Method predict()

Predict for a data set. Additonally return the true values of the response and, if the benchmark_col attribute of the Data object is not NULL, the values of the benchmark.

Usage

Model$predict(data, quiet = FALSE)

Arguments

data

Data object. Specifications on the data. Read it in if needed.

quiet

logical. Whether to suppress messages. Default is FALSE.

Returns

A list of three named numeric vectors named "predictions", "actual", and "benchmark".

Method clone()

The objects of this class are cloneable with this method.

Usage

Model$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.