Nested cross-validation for second-stage OOB predictions

Perform a nested cross-validation for a late-integration scheme, i.e., perform a cross-validation for the early model and then train second-stage models on cross-validated predictions and evaluate the entire models via out-of-bag (OOB) predictions of the second-stage models (usually a random forest).

Usage

nested_cv_oob(
  x,
  y,
  fitter1,
  fitter2,
  hyperparams1,
  hyperparams2,
  n_folds = 10,
  pseudo_cv = TRUE
)

Arguments

x

A numeric matrix holding the predictor features: rows are samples and columns are features.

y

A numeric vector holding the response variable.

fitter1

A function that fits the early model. It must be able to perform a cross validation and return an S3 class with

a cv_predict attribute, a list of vectors, for every lambda the cross-validated predictions.

a lambda attribute, a numeric vector holding the lambda values, and

a predict method.

fitter2

A function that fits the late model. It must return an S3 object with a predict method.

hyperparams1

A named list with hyperpamaters we will pass to fitter1 to obtain cross-validated predictions from the early model.

hyperparams2

A named list with hyperparameters for the late model. Unlike hyperparams1, we call fitter2 for every combination of values in hyperparams2 and lambda value from fitter1.

n_folds

An integer specifying the number of folds for the cross-validation.

pseudo_cv

A logical indicating whether we use pseudo cross-validation. I.e., fitter1 performs one n_folds cross validation, get a new training set by replacing fitter1's input features by the cross-validated predictions and procede with model fitting and getting out-of-bag predictions. This a is not real independent testing since the OOB prediction for a given sample according to fitter2 involves a model that saw the sample during training. Yet, this pseudo method saves us a factor of n_folds in computational time compared to a classical nested cross-validation. The latter will follow in future versions soon.

Value

An S3 object with class nested_fit, the model with the best performance according to the out-of-bag (OOB) predictions based on cross-validated predictions from the early model.

Details

This function does hyperparameter tuning for a nested model, i.e., a so-called early model makes predictions from the high-dimensional part of data (e.g. RNA-seq, Nanostring). We then provide these predictions as a one-dimensional continuous feature together with new features to a late model that enables a performance estimate via out-of-bag (OOB) predictions (typically a random forest).