Prepend a prefix to the file attribute of a list of AssScalar objects

Modify the AssScalar objects in place. This function supports the paradigm of specifying the Models for for all data sets at one place and them quickly adopting for a specific data set.

Usage

prepend_to_filename(ass_scalar_list, prefix)

Arguments

ass_scalar_list

list of AssScalar objects.

prefix

character Prepend this prefix to the file attribute of each element in ass_scalar_list.

Value

list of modified AssScalars.