Wrap ranger::ranger() into a patroklos-compliant fitter

This function is a patroklos-compliant fitter with validated predictions, it has a return value with a val_predict attribute.

Usage

ptk_ranger(
  x,
  y_bin,
  y_cox,
  rel_mtry,
  mtry = NULL,
  skip_on_invalid_input = FALSE,
  ...
)

Arguments

x

Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.

y_bin

Named numeric matrix with one column. Binary response.

y_cox

Named numeric matrix with two columns. The first column is the time to event, the second column indicates censoring (1 for event, 0 for censored).

rel_mtry

logical. If TRUE, interprete mtry as relative to sqrt(ncol(x)) (the recommended value), rounded to the next integer. Otherwise, take mtry directly.

mtry

Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.

skip_on_invalid_input

Logical. If TRUE and invalid input is detected, return NA instead of an error. This is useful when calling this function from inside long_nestor().

...

Further arguments passed to the wrapped function.

Value

A ptk_ranger S3 object, a ranger S3 object with the (OOB) predictions attribute renamed to val_predict.