Write data info to a JSON file

Automatically generate info about the data and write it to a JSON file. If the file already exists, write the automatically inferable info into the file. If the file is not present yet, create a skeleton, write the automatically inferable info into it. You can then fill the empty fields in by writing the json by hand. Manually already filled out info will not be overwritten.

Usage

write_data_info(filename, data, expr_tbl)

Arguments

filename

string. The path to the JSON file.

data

Data object. Analyzed data.

expr_tbl

tibble. The expression data. Rows are genes, columns are samples. The first column, named data$gene_id_col, holds the gene identifiers.

Value

The JSON string that was written to the file.