An R6 class to assess a model with one metric versus another.

Assess how well a model can predict time to event less than a certain threshold with two metrics (like the ROC-AUC and the Brier score) to guide thresholding the continuous to binary model output.

Public fields

x_metric, y_metric

Metrics shown on both axes.

pivot_time_cutoff

Assess for predicting time to event less than pivot_time_cutoff

benchmark

Incorporate this benchmark (from pheno data) into assessment.

file

Store results in this file.

lambda

Assess model with this regularization parameter.

ci_level

For any confidence intervals calculated use this level.

fellow_csv

Whether to also store the plotted data in a csv including cutoffs.

scores_plot

Whether to additionally plot rank versus model output score.

show_plots

Whether to print all generated plots.

title

Plot title.

x_lab, y_lab

Axis labels.

xlim, ylim

Axis limits.

smooth_method

Method to smooth the plot.

smooth_benchmark

Whether to also smooth the benchmark.

smooth_se

Whether to add standard error bands to the smoothed lines.

scale_x, scale_y

The scale of the axes.

vline, hline

Vertical/horizontal lines to be added to the plot.

text_size

The size of in-plot text.

text

Text label added to the plot.

alpha

The alpha value for the points in the plot.

colors

Colors used by ggplot2.

theme

Theme applied to the plot.

width, height, units

The size of the plot in units.

dpi

Plot resolution in dots per inch.

data

The data underlying the plots.

Active bindings

x_metric, y_metric

Metrics shown on both axes.

x_lab, y_lab

Axis labels.

xlim, ylim

Axis limits.

scale_x, scale_y

The scale of the axes.

vline, hline

Vertical/horizontal lines to be added to the plot.

width, height, units

The size of the plot in units.

Methods

Public methods

Ass2d$new()

Ass2d$assess()

Ass2d$plot_risk_scores()

Ass2d$assess_center()

Ass2d$clone()

Method new()

Create a new Ass2d instance.

Usage

Ass2d$new(
  x_metric,
  y_metric,
  pivot_time_cutoff,
  lambda = "lambda.min",
  benchmark = NULL,
  file = NULL,
  ci_level = 0.95,
  fellow_csv = FALSE,
  scores_plot = FALSE,
  show_plots = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf),
  smooth_method = NULL,
  smooth_benchmark = FALSE,
  smooth_se = FALSE,
  scale_x = "identity",
  scale_y = "identity",
  vline = NULL,
  hline = NULL,
  text_size = 2.5,
  text = NULL,
  alpha = 0.5,
  width = 7,
  height = 4,
  colors = NULL,
  theme = NULL,
  units = "in",
  dpi = 300
)

Arguments

x_metric

string. The name of the performance measure to be plotted on the x-axis. All measures that can be passed to the x.measure parameter of ROCR::performance() are valid.

y_metric

string. The name of the performance measure to be plotted on the y-axis.

All measures that can be passed to the measure parameter of ROCR::performance() are valid.

"logrank" for the p-values of a logrank test. In this case, x_metric must be "prevalence" or "rpp".

"precision_ci" for the lower bound of the ci_level (see below) confidence interval of the precision according to Clopper-Pearson.

pivot_time_cutoff

numeric. Time-to-event threshold that divides samples into a high/low-risk (time to event below/above pivot_time_cutoff) group. Model performance will be measured in terms of how well it can separate these two groups as a binary classifier.

lambda

numeric or string. The regularization parameter of the model to evaluate.

benchmark

string. Incorporate this benchmark (from pheno data) into the assessment if benchmark is not NULL.

file

string. If not NULL, store the resulting plot in this file.

ci_level

numeric in [0, 1]. The level used to calculate confidence intervals.

fellow_csv

logical. Whether to also store the plotted data in a csv including cutoffs.

scores_plot

logical. Display the ordered scores output by the model in a scatter plot.

show_plots

logical. Whether to show the plots after creation in an interactive graphics device.

title

string. The title of the plot. Default is NULL, i.e. no title.

x_lab, y_lab

string. Axis labels. Default is x_metric and y_metric, respectively.

xlim, ylim

numeric vector of length 2. The limits for both axes. Default is no contraints.

smooth_method

string or function. Smooth method to plot an additional smoothed graph. If NULL (the default), no smoothing. Else we pass smooth_method as the method parameter to ggplot2::geom_smooth().

smooth_benchmark

logical. Whether to also smooth the benchmark data.

smooth_se

logical. Whether to add standard error bands to the smoothed lines.

scale_x, scale_y

string or transformation object (see scales::trans_new for the latter). The scale of the axes, we will pass them to the trans paramter of ggplot2::scale_x_continuous(), ggplot2::scale_y_continuous(), respectively.

vline, hline

list or NULL. Vertical/horizontal lines to be added to the plot. A list holding the arguments to pass to ggplot2::geom_vline() and ggplot2::geom_hline(), respectively. Default is NULL, no line.

text_size

numeric. The size of text passed to the size parameter of ggplot2::geom_text(). The reason this attribute exists is you might want to have text size different the global text size of the plot. See vignette("ggplot2-specs", package = "ggplot2") for more details.

text

list or NULL. Text label added to the plot. A list holding the arguments to pass to ggplot2::geom_label(). Default is NULL, no text at all.

alpha

numeric in [0, 1]. The alpha value for the points and lines in the plot.

width

numeric. The width of the plot in units.

height

numeric. The height of the plot in units.

colors

character vector. The colors to be used for the different models. Default is NULL, which means that the default colors of ggplot2 will be used.

theme

S3 object inheriting from "theme" and "gg" (typically the return value of ggplot2::theme() or a complete ggplot2 theme like ggplot2::theme_light()). The theme of the plot. Default is NULL, which means the default theme of ggplot2.

units

string. The units of width and height. Default is "in" (inches).

dpi

numeric. Plot resolution in dots per inch.

Returns

A new Ass2d object.

Method assess()

Assess a single model (with multiple splits) on a data set.

Usage

Ass2d$assess(data, model, quiet = FALSE, msg_prefix = "")

Arguments

data

Data object. Assess on this data. The cohort attribute must be set.

model

Model object. Assess this model, multiple splits are supported.

quiet

logical. Whether to suppress messages.

msg_prefix

string. Prefix for messages. Default is "".

Details

We add the data underlying the plots to Ass2d object as a new attribute named data.

Method plot_risk_scores()

Plot rank versus model score on a data set. You can do this on the fly, i.e. you actually initialize the Ass2d object for another purpose and this method will infer reasonable defaults for the plot.

Usage

Ass2d$plot_risk_scores(data, model, quiet = FALSE, msg_prefix = "")

Arguments

data

Data object. Plot the scores for this data.

model

Model object. Plot the scores for this model.

quiet

logical. Whether to suppress messages.

msg_prefix

string. Prefix for messages.

Method assess_center()

Assess multiple models (with multiple splits) on a data set. This is a wrapper around assess() and optionally plot_risk_scores() for multiple models.

Usage

Ass2d$assess_center(
  data,
  model_list,
  model_tree_mirror = c("models", "results"),
  risk_scores = FALSE,
  comparison_plot = FALSE,
  quiet = FALSE
)

Arguments

data

Data object. Assess on this data set.

model_list

list of Model objects. Assess these models. We infer the AssSpec2d for the single plots in a reasonable way from it.

model_tree_mirror

character vector of length 2. Store the single plots in the same directory as the corresponding model for the train cohort and the comparison plot at file. When doing the analogous for the test data, replace the first element of model_tree_mirror by the second element in every file path.

risk_scores

logical. Whether to also plot rank versus risk score along with the assessment "on-the-fly".

comparison_plot

logical. Whether to generate a plot holding the assessment of all models. Default is TRUE.

quiet

logical. Whether to suppress messages. Default is FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

Ass2d$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.