An R6 class for the data

A Data object holds the phenotype and expression data belonging to a data set. It specifies the names of important features in the columns, reads from and writes to csv files and takes care of splitting the data into train and test cohort.

Public fields

name

How to refer to the data in plots and messages.

directory

Directory where the expression and pheno csv files lie.

train_prop

Proportion of samples in the training cohort.

expr_mat

The expression matrix.

pheno_tbl

The phenotype tibble.

expr_file

Name of the expression csv file inside directory.

pheno_file

Name of the pheno data csv inside directory.

cohort

The cohort of the data, train or test.

patient_id_col

The name of the column in the pheno data that holds the patient identifiers.

time_to_event_col

The name of the column in the pheno data that holds the time-to-event values.

event_col

The name of the column in the pheno data that holds the event status encoded as 1 = occurrence, 0 = censoring.

benchmark_col

The name of the column in the pheno data that holds the benchmark risk score (like the IPI).

gene_id_col

The name of the column in the expression data that holds the gene identifiers.

split_col_prefix

Column-name prefix for those columns holding splits into train and test cohort.

pivot_time_cutoff

Pivotal time cutoff for the analysis and, explicitly, for splitting the data into the train and test cohort.

imputer

Function handling NAs in the predictor matrix.

Methods

Public methods

Data$new()

Data$read()

Data$prepare()

Data$survival_quantiles()

Data$qc_preprocess()

Data$clone()

Method new()

A Data object specifies the location and format of the expression and pheno data of a single data set. This enables reading and preparing the data.

Usage

Data$new(
  name,
  directory,
  train_prop,
  pivot_time_cutoff,
  time_to_event_col,
  event_col = "progression",
  benchmark_col = NULL,
  cohort = NULL,
  expr_file = "expr.csv",
  pheno_file = "pheno.csv",
  patient_id_col = "patient_id",
  gene_id_col = "gene_id",
  split_col_prefix = "split_",
  imputer = mean_impute
)

Arguments

name

string. A telling name for the data set (e.g. "Schmitz et al. 2018").

directory

string. The directory where both expression and pheno csv files lie.

train_prop

numeric in (0, 1). The proportion of samples to draw for the training cohort in every splitting run.

pivot_time_cutoff

numeric. Pivotal time cutoff for the analysis and, explicitly, for splitting the data into the train and test cohort: If a numeric in (0, 1), preserve the the proportion of individuals below and above time_cutoff in both cohorts. Default is NULL, i.e. no further constraints on splitting.

time_to_event_col

string. The name of the column in the pheno data that holds the time-to-event values.

event_col

string. The name of the column in the pheno data that holds the event status encoded as 1 = occurrence, 0 = censoring.

benchmark_col

string or NULL. The name of the column in the pheno data that holds the benchmark risk score (like the IPI).

cohort

string in c("train", "test") or NULL. The cohort, train or test, to prepare the data for. If NULL, the default, some functions will set it themselves (e.g. training_camp() to "train", assess_2d() to "test"), others will throw an error.

expr_file

string. The name of the expression csv file inside directory. Default is "expr.csv". See details for the expected format.

pheno_file

string. The name of the pheno data csv inside directory. Default is "pheno.csv". See details for the expected format.

patient_id_col

string. The name of the column in the pheno data that holds the patient identifiers.

gene_id_col

string. The name of the column in the expression data that holds the gene identifiers.

split_col_prefix

string. Column-name prefix for those columns holding splits into train and test cohort. Some of these columns may be present in the pheno data already, others will be added during the training process if required. Default is "split_", i.e., split columns are named "split_1", "split_2", etc.

imputer

function or NULL. Function handling NAs in the predictor matrix. It takes a single argument x, the predictor matrix, a numeric with categorical variables dichotomized to binary dummy variables, and returns a matrix of the same shape with non-NAs left untouched. Default is mean_impute(). NULL means no imputation (or applying the identity function on x).

Details

The pheno csv file holds the samples as rows (with unique sample ids in the first (character) column called patient_id_col), the variables as columns.

The expression csv file holds the genes as rows (with unique gene ids in the first (character) column called gene_id_col), the samples as columns.

While the computational representation of both expression and pheno data will change over the course of the pipeline, a Data object will hold timeless information on the data.

Returns

A Data object.

Method read()

Read expression data into the matrix expr_mat and pheno data into the tibble pheno_tbl. Both will hold patients as rows.

Usage

Data$read()

Method prepare()

Prepare the read-in data for a model.

Usage

Data$prepare(model, quiet = FALSE)

Arguments

model

A Model object.

quiet

logical. If TRUE, suppress messages.

Details

You need to set the cohort before calling this method.

Method survival_quantiles()

Calculate the quantiles of the survival times.

Usage

Data$survival_quantiles(round_digits = 3)

Arguments

round_digits

integer. Round the numbers in the returned tibble to this number of digits after the point.

Details

We take censoring into account.

Returns

A tibble with two columns. For each quantile q in the first column, the time-to-event value in the second column.

Method qc_preprocess()

Quality control at the end of preprocessing

Usage

Data$qc_preprocess(expr_tbl)

Arguments

expr_tbl

A tibble with the expression data, the product of reading in an expression csv file as described in the details of the initialize() method.

Details

Check if the expression and pheno tibble are consistent with the other attributes of the Data object. You typically call this method at the end of preprocessing, and the read() method calls it.

Method clone()

The objects of this class are cloneable with this method.

Usage

Data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.